add: working my account view; refactor: move UserProfile to own component since now used in 2 views; change user links and query key to SUB instead of ID; NOTE: still hacky since API does not support getting a single user by sub
