+++ src/lib/api/user.ts
@@ -45,16 +45,16 @@ export async function deleteUser(id): Promise<boolean> {
 }
 
 /** THIS WORKS DUE TO A HACK that compensates for the API not accepting user_id key in the request body */
-export async function getUser(userId: number): Promise<User | null> {
+export async function getUser(sub: string): Promise<User | null> {
   const response = await API.post({
     apiName: "auth",
     path: "/getUsers",
-    options: { body: { user_id: userId } },
+    options: { body: { sub } },
   }).response;
   const usersResponse = (await response.body.json()) as GetUsersAPIResponse;
   // const [user] = usersResponse;
   // HACK UNTIL API WORKS TO GET A SINGLE USER
-  const user = usersResponse.find((u) => u.user_id === userId) ?? null;
+  const user = usersResponse.find((u) => u.sub === sub) ?? null;
   return user;
 }
