+++ src/lib/components/index.ts
@@ -8,3 +8,4 @@ export { default as StripedListRow } from './StripedListRow'
 export { default as TableFilter } from './TableFilter'
 export { default as TablePager } from './TablePager'
 export { default as Timer } from './Timer'
+export { default as UserProfile } from './UserProfile'
