+++ src/routeTree.gen.ts
@@ -17,6 +17,7 @@ import { Route as IndexImport } from './routes/index'
 import { Route as AuthVisitsImport } from './routes/_auth.visits'
 import { Route as AuthNewVisitImport } from './routes/_auth.new-visit'
 import { Route as AuthNewNotificationImport } from './routes/_auth.new-notification'
+import { Route as AuthMeImport } from './routes/_auth.me'
 import { Route as AuthGuestsImport } from './routes/_auth.guests'
 import { Route as AuthAdminImport } from './routes/_auth._admin'
 import { Route as AuthServicesServiceIdImport } from './routes/_auth.services_.$serviceId'
@@ -59,6 +60,12 @@ const AuthNewVisitRoute = AuthNewVisitImport.update({
 const AuthNewNotificationRoute = AuthNewNotificationImport.update({
   id: '/new-notification',
   path: '/new-notification',
+  getParentRoute: () => AuthRoute,
+} as any)
+
+const AuthMeRoute = AuthMeImport.update({
+  id: '/me',
+  path: '/me',
   getParentRoute: () => AuthRoute,
 } as any)
 
@@ -142,6 +149,13 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof AuthGuestsImport
       parentRoute: typeof AuthImport
     }
+    '/_auth/me': {
+      id: '/_auth/me'
+      path: '/me'
+      fullPath: '/me'
+      preLoaderRoute: typeof AuthMeImport
+      parentRoute: typeof AuthImport
+    }
     '/_auth/new-notification': {
       id: '/_auth/new-notification'
       path: '/new-notification'
@@ -222,6 +236,7 @@ const AuthAdminRouteWithChildren = AuthAdminRoute._addFileChildren(
 interface AuthRouteChildren {
   AuthAdminRoute: typeof AuthAdminRouteWithChildren
   AuthGuestsRoute: typeof AuthGuestsRoute
+  AuthMeRoute: typeof AuthMeRoute
   AuthNewNotificationRoute: typeof AuthNewNotificationRoute
   AuthNewVisitRoute: typeof AuthNewVisitRoute
   AuthVisitsRoute: typeof AuthVisitsRoute
@@ -232,6 +247,7 @@ interface AuthRouteChildren {
 const AuthRouteChildren: AuthRouteChildren = {
   AuthAdminRoute: AuthAdminRouteWithChildren,
   AuthGuestsRoute: AuthGuestsRoute,
+  AuthMeRoute: AuthMeRoute,
   AuthNewNotificationRoute: AuthNewNotificationRoute,
   AuthNewVisitRoute: AuthNewVisitRoute,
   AuthVisitsRoute: AuthVisitsRoute,
@@ -246,6 +262,7 @@ export interface FileRoutesByFullPath {
   '': typeof AuthAdminRouteWithChildren
   '/login': typeof LoginRoute
   '/guests': typeof AuthGuestsRoute
+  '/me': typeof AuthMeRoute
   '/new-notification': typeof AuthNewNotificationRoute
   '/new-visit': typeof AuthNewVisitRoute
   '/visits': typeof AuthVisitsRoute
@@ -261,6 +278,7 @@ export interface FileRoutesByTo {
   '': typeof AuthAdminRouteWithChildren
   '/login': typeof LoginRoute
   '/guests': typeof AuthGuestsRoute
+  '/me': typeof AuthMeRoute
   '/new-notification': typeof AuthNewNotificationRoute
   '/new-visit': typeof AuthNewVisitRoute
   '/visits': typeof AuthVisitsRoute
@@ -278,6 +296,7 @@ export interface FileRoutesById {
   '/login': typeof LoginRoute
   '/_auth/_admin': typeof AuthAdminRouteWithChildren
   '/_auth/guests': typeof AuthGuestsRoute
+  '/_auth/me': typeof AuthMeRoute
   '/_auth/new-notification': typeof AuthNewNotificationRoute
   '/_auth/new-visit': typeof AuthNewVisitRoute
   '/_auth/visits': typeof AuthVisitsRoute
@@ -295,6 +314,7 @@ export interface FileRouteTypes {
     | ''
     | '/login'
     | '/guests'
+    | '/me'
     | '/new-notification'
     | '/new-visit'
     | '/visits'
@@ -309,6 +329,7 @@ export interface FileRouteTypes {
     | ''
     | '/login'
     | '/guests'
+    | '/me'
     | '/new-notification'
     | '/new-visit'
     | '/visits'
@@ -324,6 +345,7 @@ export interface FileRouteTypes {
     | '/login'
     | '/_auth/_admin'
     | '/_auth/guests'
+    | '/_auth/me'
     | '/_auth/new-notification'
     | '/_auth/new-visit'
     | '/_auth/visits'
@@ -370,6 +392,7 @@ export const routeTree = rootRoute
       "children": [
         "/_auth/_admin",
         "/_auth/guests",
+        "/_auth/me",
         "/_auth/new-notification",
         "/_auth/new-visit",
         "/_auth/visits",
@@ -391,6 +414,10 @@ export const routeTree = rootRoute
     },
     "/_auth/guests": {
       "filePath": "_auth.guests.tsx",
+      "parent": "/_auth"
+    },
+    "/_auth/me": {
+      "filePath": "_auth.me.tsx",
       "parent": "/_auth"
     },
     "/_auth/new-notification": {
